clean up, rm angles from generic store
